django_test/apps/feeds/serializers.py
from models import Feed, Article
from rest_framework import serializers


class FeedSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Feed
        fields = ('name', 'url')


class ArticleSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Article
        fields = ('title', 'url', 'description', 'feed')
