requirements.txt
Django==1.7
Fabric==1.8.2
#South==0.8.4
argparse==1.2.1
ecdsa==0.10
paramiko==1.12.2
pycrypto==2.6.1
wsgiref==0.1.2
elasticsearch
djangorestframework
